# static/openapi.yaml
openapi: 3.0.3
info:
  title: Task Logger API
  version: "1.0.0"
  description: >
    API for logging Start/End task events (used by an Apple Watch Shortcut
    and a web frontend) and managing a day planner. All endpoints require
    an API key, supplied either as a query param (?key=...) or a header
    (X-API-Key).

servers:
  - url: https://applewatchtriggers.fly.dev
    description: Production (Fly.io)
  - url: http://localhost:8080
    description: Local development

security:
  - ApiKeyHeader: []
  - ApiKeyQuery: []

paths:
  /toggle:
    get:
      summary: Log a Start or End event
      description: >
        Toggles based on current state: if no task is in progress, this
        starts one. If a task is already in progress, this ends it and
        calculates the duration. Backed by tasks.db (SQLite) - each Start
        creates one task row with a stable id; the matching End fills in
        that same row's end/duration_minutes.
      tags: [Time Log]
      parameters:
        - name: name
          in: query
          required: false
          description: >
            Optional task name. Only ever sent by the webpage's own
            Start/End button - the Apple Watch Shortcut never sends this,
            so every Watch-logged task gets a generic
            "Task – <weekday>, <time>" title instead. Query param (not a
            JSON body) so /toggle can stay a plain GET the Shortcut can call.
          schema: { type: string, example: "Deep work" }
      responses:
        "200":
          description: Event logged successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
                  action: { type: string, enum: [Start, End] }
                  id: { type: integer, example: 42, description: "The task's id in tasks.db" }
                  message: { type: string, example: "Task started at 09:00" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"

  /status:
    get:
      summary: Check what the next /toggle call will do
      description: Read-only - does not log anything.
      tags: [Time Log]
      responses:
        "200":
          description: Next action
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  next_action: { type: string, enum: [Start, End] }
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/logs:
    get:
      summary: Get every logged task
      description: >
        Returns every task, newest first - everything a bento-card grid
        needs to render. A currently in-progress task has "end" and
        "duration_minutes" as null. Kept under the "sessions" key for
        backward compatibility with the webpage's existing Time Log view.
      tags: [Time Log]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  sessions:
                    type: array
                    items: { $ref: "#/components/schemas/Task" }
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/logs/{task_id}:
    get:
      summary: Get full detail for one task
      description: What the task detail page loads.
      tags: [Time Log]
      parameters:
        - name: task_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Task detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  task: { $ref: "#/components/schemas/Task" }
        "404":
          description: No task with that id
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      summary: Edit a task's name/start/end
      description: >
        Send any subset of name/start/end to update. If start or end
        changes, duration_minutes is recalculated server-side so it can
        never go stale.
      tags: [Time Log]
      parameters:
        - name: task_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, example: "Deep work" }
                start: { type: string, example: "2026-09-14 09:00:00", description: "YYYY-MM-DD HH:MM:SS" }
                end: { type: string, example: "2026-09-14 09:45:00", description: "YYYY-MM-DD HH:MM:SS" }
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  task: { $ref: "#/components/schemas/Task" }
        "400":
          description: Nothing to update, or start/end isn't in YYYY-MM-DD HH:MM:SS format
        "404":
          description: No task with that id
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      summary: Delete a task
      description: Permanent - ids are never reused, so an old bento-card link never ends up pointing at a different task.
      tags: [Time Log]
      parameters:
        - name: task_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted successfully
        "404":
          description: No task with that id
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/planner:
    get:
      summary: Get planner items for a date
      tags: [Planner]
      parameters:
        - name: date
          in: query
          required: false
          description: "Format: YYYY-MM-DD. Defaults to today."
          schema: { type: string, format: date }
      responses:
        "200":
          description: Planner items for that date
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  date: { type: string, format: date }
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/PlannerItem" }
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      summary: Add a new planner item
      tags: [Planner]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [date, start_time, end_time, title]
              properties:
                date: { type: string, format: date, example: "2026-09-02" }
                start_time: { type: string, example: "09:00" }
                end_time: { type: string, example: "10:00" }
                title: { type: string, example: "Deep work" }
      responses:
        "201":
          description: Item created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  item: { $ref: "#/components/schemas/PlannerItem" }
        "400":
          description: Missing required fields
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/planner/{item_id}:
    patch:
      summary: Update a planner item
      description: Send any subset of fields to update (e.g. just "completed" to check it off).
      tags: [Planner]
      parameters:
        - name: item_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date: { type: string, format: date }
                start_time: { type: string }
                end_time: { type: string }
                title: { type: string }
                completed: { type: string, enum: ["True", "False"] }
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  item: { $ref: "#/components/schemas/PlannerItem" }
        "404":
          description: No planner item with that id
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      summary: Delete a planner item
      tags: [Planner]
      parameters:
        - name: item_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted successfully
        "404":
          description: No planner item with that id
        "401":
          $ref: "#/components/responses/Unauthorized"

components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
    ApiKeyQuery:
      type: apiKey
      in: query
      name: key

  schemas:
    Task:
      type: object
      properties:
        id: { type: integer, example: 42 }
        name:
          type: string
          example: "Deep work"
          description: >
            The name set from the webpage, or - if it was logged from the
            Watch, or from the webpage without a name - a generic
            "Task – <weekday>, <time>" title, e.g. "Task – Fri, 9:00 AM".
        date: { type: string, format: date, example: "2026-09-14" }
        start: { type: string, example: "2026-09-14 09:00:00" }
        end: { type: string, example: "2026-09-14 09:45:00", nullable: true }
        duration_minutes: { type: number, nullable: true, example: 45.0 }

    PlannerItem:
      type: object
      properties:
        id: { type: string, example: "e97f7cd2" }
        date: { type: string, format: date }
        start_time: { type: string, example: "09:00" }
        end_time: { type: string, example: "10:00" }
        title: { type: string }
        completed: { type: string, enum: ["True", "False"] }

  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            type: object
            properties:
              ok: { type: boolean, example: false }
              message: { type: string, example: "Invalid or missing API key" }
    ServerError:
      description: Server-side error (e.g. file read/write failure)
      content:
        application/json:
          schema:
            type: object
            properties:
              ok: { type: boolean, example: false }
              message: { type: string }